# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MediTrack Pro API — PULSO IoT Pill Dispensing System
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /patients:
    get:
      operationId: listPatients
      tags: [patients]
      summary: List patients linked to current caregiver
      responses:
        "200":
          description: List of patients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Patient"
    post:
      operationId: createPatient
      tags: [patients]
      summary: Create a patient profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatientInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
  /patients/{patientId}:
    get:
      operationId: getPatient
      tags: [patients]
      summary: Get a patient by ID
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Patient
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updatePatient
      tags: [patients]
      summary: Update a patient profile
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatientUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"

  # ── Caregivers ─────────────────────────────────────────────
  /patients/{patientId}/caregivers:
    get:
      operationId: listPatientCaregivers
      tags: [caregivers]
      summary: List caregivers linked to a patient
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of caregivers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CaregiverLink"
    post:
      operationId: inviteCaregiver
      tags: [caregivers]
      summary: Invite a caregiver to a patient's care circle
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CaregiverInviteInput"
      responses:
        "201":
          description: Invited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaregiverLink"
  /patients/{patientId}/caregivers/{caregiverId}:
    delete:
      operationId: removeCaregiver
      tags: [caregivers]
      summary: Remove a caregiver from the care circle
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: integer
        - name: caregiverId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ── Medications ────────────────────────────────────────────
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    # Auth
...
    PatientInput:
      type: object
      required: [name, dateOfBirth, bloodType]
      properties:
        name:
          type: string
        dateOfBirth:
          type: string
          format: date
        bloodType:
          type: string
          enum: [A+, A-, B+, B-, AB+, AB-, O+, O-]
        conditions:
          type: array
          items:
            type: string
        allergies:
          type: array
          items:
            type: string
        emergencyContactName:
          type: string
        emergencyContactPhone:
          type: string
        preferredLanguage:
          type: string
          enum: [en, fil]
          default: en
    PatientUpdate:
      type: object
      properties:
        name:
          type: string
        dateOfBirth:
          type: string
          format: date
        bloodType:
          type: string
          enum: [A+, A-, B+, B-, AB+, AB-, O+, O-]
        conditions:
          type: array
          items:
            type: string
        allergies:
          type: array
          items:
            type: string
        emergencyContactName:
          type: string
        emergencyContactPhone:
          type: string
        preferredLanguage:
          type: string
          enum: [en, fil]
    Patient:
      type: object
      required: [id, name, dateOfBirth, bloodType, conditions, allergies, emergencyToken]
      properties:
        id:
          type: integer
        name:
          type: string
        dateOfBirth:
          type: string
          format: date
        bloodType:
          type: string
        conditions:
          type: array
          items:
            type: string
        allergies:
          type: array
          items:
            type: string
        emergencyContactName:
          type: ["string", "null"]
        emergencyContactPhone:
          type: ["string", "null"]
        emergencyToken:
          type: string
        preferredLanguage:
          type: string
        createdAt:
          type: string
          format: date-time

    # Caregiver
    CaregiverInviteInput:
      type: object
      required: [email, accessLevel]
      properties:
        email:
          type: string
        accessLevel:
          type: string
          enum: [view, full]
    CaregiverLink:
      type: object
      required: [id, caregiverId, patientId, accessLevel, status]
      properties:
        id:
          type: integer
        caregiverId:
          type: integer
        patientId:
          type: integer
        caregiverName:
          type: string
        caregiverEmail:
          type: string
        accessLevel:
          type: string
          enum: [view, full]
        status:
          type: string
          enum: [pending, active]
        createdAt:
          type: string
          format: date-time

    # Medication
...